Boid behaviour and Obstacle Avoidance
- LookAt(rb.velocity) macht keinen sinn! -> besser ist LookAt(transform.position + rb.velocity) (muss immer abhängig von der aktuellen position gemacht werden)
- Probleme mit der Gewichtung von Boid Verhalten (Wenn Seperation höher ist als cohesion sammeln sie sich nicht mehr)
- eine Sphere aus Raycasts zu machen ist simpel, einfach ein quadrat (3x for-lopp) und die Länger der Rays normalizen

+++ Assets/Boids/Scripts/BoidSettings.cs
@@ -9,7 +9,7 @@ public class BoidSettings : ScriptableObject
     [SerializeField] private float radius;
     [SerializeField] private LayerMask boidLayer;
     [SerializeField] private float speed;
-    [SerializeField] private int avoidanceRadius;
+    [SerializeField] private float avoidanceRadius;
     [SerializeField] private LayerMask obstacleLayer;
     [SerializeField] private float avoidanceSensitivity;
 
@@ -19,7 +19,7 @@ public class BoidSettings : ScriptableObject
     public float Radius => radius;
     public LayerMask BoidLayer => boidLayer;
     public float Speed => speed;
-    public int AvoidanceRadius => avoidanceRadius;
+    public float AvoidanceRadius => avoidanceRadius;
     public LayerMask ObstacleLayer => obstacleLayer;
     public float AvoidanceSensitivity => avoidanceSensitivity;
 }
